persistent-vars-storage-bcm2711: adds a Raspberry Pi compatible persistent storage utility
The Morse Micro persistent-vars-storage package is used to load and store certain device
parameters or configuration from non-volatile storage on a device. This is useful for a
more persistent configuration than what is offered by UCI. Typically used for dpp keys.

The bcm2711 version targets Raspberry Pis and pushes this configuration to the on-board
eeprom.