persistent-vars-storage-ubootenv: adds a uboot-env compatible persistent storage utility
The Morse Micro persistent-vars-storage package is used to load and store certain device
parameters or configuration from non-volatile storage on a device. This is useful for a
more persistent configuration than what is offered by UCI. Typically used for dpp keys.

The uboot-env version targets any device which can allocate a uboot-env style partition
in flash dedicated for this purpose. It is recommended to not use the actual uboot-env
partition, but rather create a new partition using the uboot-env format.