morse-bundle: adds a package deploying common utilities
The morse-bundle package is used to deploy common utilities, scripts,
and configuration which otherwise have no dedicated home.

While not an ideal approach to achieving the above, it does ensure the
evaluation kits can provide an identical user experience without impacting
customers who do not require some of the changes.