morse_mesh11sd: add the Morse Micro modified mesh11sh package
mesh11sd is used by OpenWrt to provide dynamic configuration for 802.11s
mesh-like networks.

Because the Morse Micro design/implementation of mesh requires configuration
changes to be made to the supplicant configuration files, rather than through
iw/netlink, the modifications in this package allow for the mesh11sd package
to reload the newly applied configuration.