luci-app-ekhwizards/morseconfig: (APP-3854, PR #1036) force writes and update diagram with defaults
There is code in the wizard to auto-populate the default wifi key
and password in appropriate contexts. Unfortunately, because
from the form's perspective it looks like the existing value,
Luci's caching smarts strike again and it doesn't write this
value out unless it changes.

Moreover, because we have multiple widgets in the wizard backing onto the
same default_radio0 or equivalent - i.e. to support ap/sta mode
on the 'same' interface, there's an SSID field for the AP side
and an SSID field for the STA side which have different
behaviour and are rendered on different pages - we can't
simply force this value in uci at the start of the wizard,
as it has to be dynamic (i.e. only set if that particular
widget ended up being used).

So, having dug this hole, we now dig deeper by attaching
forcewrite to all these dodgy loads, AND forcing the uci
to update immediately so the diagram will render correctly
as soon as the fake 'page' is loaded.

Approved-by: Andrew Dodd
Approved-by: Evan Benn