luci-app-ekhwizards: (APP-4041, PR #25) revert 2.4GHz radio changes in PR #2
This reverts commit 71484fba2870c0faeaad33d3760bfd7915b6c4ab.

Why?

(a) this commit breaks the non prplmesh wizards (due to the way it mutates
    any wifi devices
(b) it should not be mutating wifi devices in resetNetworkTopology
(c) we don't have testing for it yet.

Suggest that before this is introduced we need to make sure the ekh03v4
default config is included in the wizard and tests results are generated.
We can then more easily understand what this is changing.


Approved-by: Evan Benn
Approved-by: Lyall Beveridge