netifd-morse: (APP-4391, PR #66) prevent bad hotplug
It's unclear _why_ we're seeing this, but apparently
in testing we somehow end up with a 'double' hotplug
where the first device didn't have a path/macaddr.
This first device is then both useless (since
it can't be identified/brought up) and confusing.

The good news is that in the only times we've seen
this there has been a good subsequent hotplug event.


Approved-by: Sophronia Koilpillai
Approved-by: Lyall Beveridge