morse_driver: (APP-4396, PR #21) fix incorrect MORSE_RC option
The option changed in the driver to be a DISABLE=y one.

Also remove attempt to double set flags (which are set
in the driver Makefile based on the config options, and here
are overwritten anyway by NOSTDINC= later).

Note that setting this won't actually work because of an issue in the driver (see SW-14492)

Approved-by: Prabha Murugesan
Approved-by: Andrew Dodd