dpp-qrcode-key/netifd-morse: (APP-4413, PR #70) move qrcode extraction into hotplug
When the hotplug happens we have enough information (key + mac address)
to generate the qrcode. Yes, more stuff is happening in one script, but
it means we don't have a complicated dependency situation. We also get
to remove this generation out of netifd (since not everyone will want it).

Note that this script is still not smart enough to handle multiple morse
devices. A subsequent hot plug with a different MAC address will regenerate
the qrcode. Note that this behaviour is useful if you have only one device,
since you may force the MAC address via the module parameters.

* dpp-qrcode-key: (APP-4413) fix hotplug file location


Approved-by: Milad Mohtashamirad
Approved-by: Lyall Beveridge