morse-leds/morse-button: (APP-4220, PR #80) centralise LED/button handling
Handle all 'morse style' LEDs in a single overrides file, with
features based on what named LEDs exist rather than what the device is.
This makes it easier for us to be flexible with various EKH devices,
and most importantly easily support arbitrary routers.

Also, we change the standard BTN_0 handler to a reset button.
This means it is more likely to work correctly on arbitrary
routers (e.g. ekh19) which already have a reset button.
Note that the artini button handler (morse-modeswitch-button)
remains on BTN_0.


Approved-by: Sophronia Koilpillai
Approved-by: Milad Mohtashamirad