luci-app-rangetest: (APP-4524, PR#90) filter the HaLowLink 1 management address (10.22.121.111) in the remote device selection dropdown
Given the increasing popularity of the HaLowLink 1 and subsequent use of the range test UI, this IP address only serves to cause confusion and will always fail with the range test.

This does pose a risk where on a standard EVK a person may be using that address and be unable to find it, but in that chance is very low given the obscurity of the management IP and the risk versus reward upside of globally reducing user frustration makes this justified.


Approved-by: Sophronia Koilpillai
Approved-by: James Haggerty