netifd-morse: (APP-4567, PR #98) make dppd properly respond to hostapd changes
Like smart manager, dppd needs to talk to hostapd, and morse.sh
is the one that know when things have changed. We could manually
push some kind of ubus trigger, but since we already have this
mechanism let's consolidate.

It appears this may have been the original intent of APP-1762


Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn