luci-app-rangetest: (APP-4454, PR #103) replace remote-rangetest-rpc with uclient fetch API methods
This change removes the LUA-based extra ubus service which was managing the outgoing calls and replaces it with a simple method in the range test backend API. This change provides no extra functionality except to consolidate code. 

All system calls in this PR have also been changed to use arrays in order to prevent shell injection attacks where user input is being parsed.


Approved-by: Sophronia Koilpillai
Approved-by: James Haggerty