APP-4574 PR #104: Enable wds_bridge and bridge wifi-iface config
There were two wds related parameters not being set by morse.sh netifd
configs. bridge= was intentionally disabled with a comment that this
made something break. wds_bridge was not working due to a missing
declaration.

This patch re-enables both of these and produces hostapd configs more
similar to mac80211.sh configs. A standard AP on a bridge seems to work
as before. Both parameters are optional according to hostapd. There is
no change to wpa_supplicant configs.

Tested by creating a WDS AP bridged with ethernet and associating a WDS
client. A device attached to the AP ethernet can communicate with
a device attached to the client ethernet.

If wds_bridge is set to a different bridge that does not exist, hostapd
still adds clients to the bridge= bridge. This may be a bug. If the
bridge does exist the clients are added to it. Traffic does not work in
either of these cases, either of these cases, so the wds_bridge
parameter only works when set the brid…

Approved-by: Louise Tan
Approved-by: James Haggerty