netifd-morse/luci-app-morseconfig: (APP-4617, PR #110) reintroduce multiple BSS support
This reintroduces the approach from APP-1598, but with some tweaks
that involved a bit more refactoring, including:

- better identification of when multiple interfaces exist
  (and only one point this happens)
- no start hostapd/kill hostapd when second hostapd arrives
- corrected cleanup code (important for multiple interfaces)

Context: this is _not_ how OpenWrt usually handles multiple BSSes,
where hostapd is configured to emit a MBSSID IE in a single beacon.
Instead, we essentially run two hostapds in the same process
which in this case target the same radio.

Approved-by: Sophronia Koilpillai
Approved-by: Lyall Beveridge