luci-app-morseconfig: (APP-4776, PR #114) forceBridge should retain existing ports
When using forceBridge if the network was currently non-bridged it
would lose the port entirely.

I've refactored this at the same time to stop changing the typing
of the 'bridge' variable (which started out as an object, but on
one code path changed to a name, and then was used later...).


Approved-by: Sophronia Koilpillai
Approved-by: Prabha Murugesan