luci-theme-morseargon: (APP-4604, PR #116) make admin wireless css table hack specific
Currently, all tables on the wireless advanced page are affected by this CSS
hack (which removes all the heading/colour info by overriding this on the cell).
This looks like it was intended for the association list and list of
wifi ifaces/devices on the main page, where the section header serves
as the title and the table colouring is too noisy, but is annoying when
you want to construct other tables.

As far as I can determine, this affects the scan list results +
the new table I'm adding for BCF info.


Approved-by: Prabha Murugesan
Approved-by: Lyall Beveridge