*: (APP-4172) aggressively remove mesh_id/ssid
It's harmless in terms of UCI to leave mesh_id/ssids around
when in a mode they don't apply to (e.g. ap/ssid vs mesh/mesh_id),
but:

- it confuses people looking at UCI files
- if you load the luci network-wireless page then save, it _will_
  try to remove the non-applicable one, causing spurious changes

Here, I only try to clean up the situation on the primary
wifi-iface (default_radio0) to cut down on the amount of code,
since it's less likely that the other ifaces will change in
strange ways (or at least those changes won't be caused by us).

At the same time, since we're acknowledging that the main
iface on an HL1 might have been changed to mesh (and hence
is more likely to have _lost_ the SSID), we need to restore the
SSID on the HL1.

With the change to wizard-config, only the removal of ssid when
setting mesh is strictly necessary, as the wizard-config
is only run as the first configuration (and is not smart enough to
remove old configuration anyway - i.e. it's tied to the firstboot
config). However, in case this code is used elsewhere or the
role of wizard-config is changed, we may as well be consistent.


Approved-by: Sophronia Koilpillai
Approved-by: Lyall Beveridge