netifd-morse/morse-bundle: (APP-4786, PR #125) always undisable morse module in wireless-defaults
Remove disabled from the morse wifi-device even if country isn’t set. In this state, it will load with the default region BUT will apply the BCF. This is useful, since we will then be able to run iwinfo countrylist so that the landing page and other pages will display the correct list of supported countries (separate PR).

To achieve this, morse.sh is cleaned up so it better understands that the module is loaded but no country has been specified by the user (since the driver will always want a country in order to successfully load). This works by using the driver default country (currently US - set in our morse_driver package), but then refusing to actually bring up any ifaces (since the user hasn’t explicitly requested a country yet).

Approved-by: Prabha Murugesan
Approved-by: Andrew Dodd
Approved-by: Evan Benn