netifd-morse: (APP-4879, PR #137) only force the wpa_supplicant mac address if fixed mac
We want to have the STA/mesh MAC address be the same as the
bridge MAC address. However, if the bridge MAC address
hasn't been fixed yet (either via UCI or because there's
a primary port), we shouldn't use it as it may change
later in the bring-up.

This happens if there are no ethernet ports on the bridge
for netifd to choose a 'primary'. However, there's
a connected patch to netifd to choose a WiFi STA/mesh
MAC address as the primary once it's attached.


Approved-by: Andrew Dodd
Approved-by: Evan Benn