morse-bundle: (APP-5013, PR #155) simplify hostname configuration
Due to the complexities introduced by the current ekh01
image names (and their changes due to hotplug), keep the existing
behaviour by simply forcing all ekh0x devices to have ekh0x as the hostname.

This means that we can remove the complexity from the default
hostname behaviour (previously it was using sysinfo_model,
which could be quite large).


Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn