morse-dppd: (APP-5065, PR #61) always pick first iface for dpp qrcode
With the current code, it was picking the second key and the
first iface, then happily putting that key into the
wpa_supplicant conf (i.e. overwriting the first key with
the second key, so that both APs had the same key).

This approach is copied from the mesh_11sd code.


Approved-by: Sophronia Koilpillai
Approved-by: Prabha Murugesan