netifd-morse: (APP-5306, PR #182) simplify channel bring-up logic
This means that 4mhz channels with a 2mhz primary use 1 instead of 2
as the default index. Arien said that this logic was
very old buildroot logic to workaround old issues
with using the first channel (since 1 would be 0+1 with a 2mhz primary).

This has the happy by-product of making Ch48 come up correctly
(as otherwise it would try to use the blocked Ch50 as the primary).


Approved-by: Prabha Murugesan
Approved-by: Evan Benn