hostapd_s1g: (APP-5117, PR #89) remove configuration hack in gas_request_handler
For OpenWrt, we restart dppd and tell it the password every time our
config changes, so this is unnecessary. More importantly, it also
removes the entire existing conf_sta, which means it removes any
custom configuration (conf_extra_name/value) we send via
dpp_custom_configuration, which is used for HL2.

For details, see SW-16596.


Approved-by: Prabha Murugesan
Approved-by: Lyall Beveridge