morse-client-config: (APP-5271, PR #201) add Extender/client config
Usually, we expected users to pair Extenders via DPP PB, which
means they never have to access the interface.

However, if they _do_ attempt to access the interface
(i.e. 192.168.12.1) previously this would simply fail.
Now we present them with an unauthenticated page which lets
them enter the credentials of the device they want to connect
to (and also lets them update the 2.4 credentials if desired).

This enables:

- users who don't want to/can't use DPP and are scared by the Quick Config
  page to configure their device
- semi-automated device provisioning (i.e. with single unauthed
  POST request)
- users to set the 2.4 credentials (incl. make them the same as the AP)

But:

- this is _unauthed_
- it's not translated, and doesn't follow our normal styling
- users are _completely_ locked out without going through this
  page -> this is intentional, for security and to make sure
  users aren't trying to custom configure extenders in strange ways
- we do _not_ do the DPP wizard-config style setup, where we
  transition to Mesh Points/EasyMesh agents etc.

Approved-by: Sophronia Koilpillai
Approved-by: Lyall Beveridge