netifd-morse: (APP-4539, PR #218) keep removing fractional sleeps
After I removed them in APP-4539, the fractional sleeps are
creeping back. I was letting them in because I thought that
FANCY_SLEEP was enabled by default, which according to the doco
allows fractional sleeps and I must have made a mistake
earlier. However, apparently even with FANCY_SLEEP
unless FLOAT_DURATION is included (which it is on our builds)
this will fail.

Approved-by: Sophronia Koilpillai
Approved-by: Milad Mohtashamirad