netifd-morse: (APP-5432, PR #214) simplify hotplug to avoid bouncing
In certain circumstances, the hotplug wifi up would trigger during
the bringup (when uci_state wasn't yet set after an rmmod/modprobe).
I've changed this so that we look at the netifd state to ignore this, and also only bring up the 'right' device.

I've also removed the disabled/disabled_by_hotplug uci entries
to keep things simple, since if the USB device was removed when
the device wasn't active we wouldn't have this behaviour anyway,
and LuCI is already able to correctly reflect devices that
are inactive because they're not plugged in.

Also catch case where morse.sh has nothing much to do, which means
it can potentially finish _before_ this hotplug runs... to make
this vanishingly unlikely, we move the relatively expensive dpp key
script to late in the hotplug, and add a sleep after insertion.

By forcing the sleep after insertion, we can also get away with
reverting to the original mac80211.sh find_phy
(since sleep 1 is surely enough for the phy to appear in
device tree, right? I know we retried for a while before...)

Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn