Merged in APP-5765-dont-bring-up-interface-in-morse-sh (pull request #265)
netifd-morse: (APP-5765, PR #265) don't bring up interface in morse.sh

* netifd-morse: (APP-5765) don't bring up interface in morse.sh

If we're planning to run hostapd/wpa_supplicant, do not bring up
the interface in morse.sh. Previously, we were inconsistent with
this depending on the interface type.

This means that the interface will not come up unless hostapd/wpa_supplicant
are correctly configured, which makes it easier for other programs
to understand what's going on.

At the same time, I've cleaned up the mac address setting so that
it happens every time, and correctly handled an iw failure on
STA bring-up (typo).


Approved-by: Milad Mohtashamirad
Approved-by: Lyall Beveridge