dpp-handler: (APP-5354, PR #271) forbid STA DPP for EasyMesh agent
This will usually only happen if the user does DPP via the CLI or via
the card, as the button will do DPP On the AP.

As part of this, improve the failure signalling and LED restore.
This means that you get the 5 sec fast flash if we can't find
a valid interface at all, which should improve the situation
for mesh Extenders.

(unfortunately, if there is no interface at the moment we don't
chase the valid LEDs down to the phy level)


Approved-by: Sophronia Koilpillai
Approved-by: Prabha Murugesan