luci-mod-home: (APP-5963, PR #307) only report mesh point connected if associated for uplink
Mesh Points (and APs) are treated as 'up' even if nothing is associated.
We should consider any wireless connection as down for the purposes of an uplink
if there is nothing in the assoclist.

Also, only report the tx/rx rate of the first thing in the assoclist
if it's a station (since otherwise there may be more than one entry
in the assoclist).


Approved-by: Prabha Murugesan
Approved-by: Lyall Beveridge