hostapd_s1g: (APP-6103, PR #137) handle bigger DPP conf objects
The DPP command (which is how we pass the DPP conf object) was unnecessarily limited.

We should now fail only when we exceed the limitations of the protocol blocks us,
which I believe is when our custom config cannot fit into a single frame
(fragmented GAS response; see 8.3 of spec).


Approved-by: Prabha Murugesan
Approved-by: Lyall Beveridge